+++ package.json
@@ -104,7 +104,8 @@
     "@typescript-eslint/eslint-plugin": "6.15.0",
     "@typescript-eslint/parser": "6.15.0",
     "@vitest/coverage-v8": "1.1.3",
-    "body-parser": "1.20.2",
+    "body-parser": "1.20.3",
+    "body-parser-v2": "npm:body-parser@2",
     "commitizen": "4.3.0",
     "cors": "2.8.5",
     "cz-conventional-changelog": "3.3.0",
@@ -115,7 +116,8 @@
     "eslint-plugin-node": "11.1.0",
     "eslint-plugin-prettier": "5.1.2",
     "eslint-plugin-tsdoc": "0.2.17",
-    "express": "4.18.2",
+    "express": "4.21.1",
+    "express-v5": "npm:express@5",
     "fastify": "4.25.2",
     "firebase-admin": "11.11.1",
     "firebase-functions": "4.5.0",
